+++ CHANGELOG.md
@@ -1,13 +1,14 @@
 Version 0.24
 ------------
-- added preliminary `curl_save()` function to download web content to local file.
-- added preliminary `dlreplacecontent()` function.
+- added preliminary `curl_save()` function to download web content to a local file.
+- added preliminary `dlreplacecontent()` function to replace the content of a linked file.
 - `dl_directory` check constraint that dirpath is like '/%/'
 - new `is_local()`, `is_valid()` as `is_http_success()` immutable functions, for use in check constraints and such
 - forbid * in file names
 - new `iri()` function to convert IRIs (unicode) to URIs (escaped) 
 - `dlvalue()` now supports `IRI` link type, bug fixes
 - dump/restore improvements (function `datalink.online()`)
+- added `datalink.http_response_codes` reference table 
 
 Version 0.23
 ------------